// src/components/training/training-lesson.enum.ts
export enum TrainingLesson {
  KEYS_F_AND_J = 'KEYS_F_AND_J',
  KEYS_D_AND_K = 'KEYS_D_AND_K',
  KEYS_FJDK = 'KEYS_FJDK',
  KEYS_S_AND_L = 'KEYS_S_AND_L',
  KEYS_Q_AND_M = 'KEYS_Q_AND_M',
  KEYS_QSLM = 'KEYS_SLQM',
  KEYS_G_AND_H = 'KEYS_G_AND_H',
  KEYS_FGHJ = 'KEYS_FGHJ',
  KEYS_A_AND_SEMI_COLON = 'KEYS_A_AND_SEMI_COLON',
  KEYS_ASL_SEMI_COLON = 'KEYS_ASL_SEMI_COLON',

  KEYS_R_AND_U = 'KEYS_R_AND_U',
  KEYS_E_AND_I = 'KEYS_E_AND_I',
  KEYS_ERUI = 'KEYS_RUEI',
  KEYS_Z_AND_O = 'KEYS_Z_AND_O',
  KEYS_A_AND_P = 'KEYS_A_AND_P',
  KEYS_AZOP = 'KEYS_AZOP',
  KEYS_T_AND_Y = 'KEYS_T_AND_Y',
  KEYS_RTYU = 'KEYS_RTYU',
  KEYS_W_AND_O = 'KEYS_W_AND_O',
  KEYS_Q_AND_P = 'KEYS_Q_AND_P',
  KEYS_QWOP = 'KEYS_QWOP',

  KEYS_V_AND_COMMA = 'KEYS_V_AND_COMMA',
  KEYS_C_AND_SEMI_COLON = 'KEYS_C_AND_SEMI_COLON',
  KEYS_CV_COMMA_SEMI_COLON = 'KEYS_CV_COMMA_SEMI_COLON',
  KEYS_X_AND_COLON = 'KEYS_X_AND_COLON',
  KEYS_W_AND_EX = 'KEYS_W_AND_EX',
  KEYS_WX_COLON_EX = 'KEYS_WX_COLON_EX',
  KEYS_B_AND_N = 'KEYS_B_AND_N',
  KEYS_VBN_COMMA = 'KEYS_VBN_COMMA',
  KEYS_V_AND_M = 'KEYS_V_AND_M',
  KEYS_C_AND_COMMA = 'KEYS_C_AND_COMMA',
  KEYS_CVM_COMMA = 'KEYS_CVM_COMMA',
  KEYS_X_AND_DOT = 'KEYS_X_AND_DOT',
  KEYS_Z_AND_SLASH = 'KEYS_Z_AND_SLASH',
  KEYS_ZX_DOT_SLASH = 'KEYS_ZX_DOT_SLASH',
  KEYS_VBNM = 'KEYS_VBNM',
}